fix: minor client bugs
