--- client/src/components/today/index.ts
@@ -1,7 +1,0 @@
-export { AIInsights } from './AIInsights';
-export { JournalActions } from './JournalActions';
-export { MoodTracker } from './MoodTracker';
-export { QuickStats } from './QuickStats';
-export { SnippetCard } from './SnippetCard';
-export { SnippetsList } from './SnippetsList';
-export { WeeklyProgress } from './WeeklyProgress';
